pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "filterworld"
version = "0.0.0"
description = "Package for visualizing computer vision model outputs"
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Matt Whiteway" },
]

keywords = ["machine learning", "deep learning", "computer vision", "visualization"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
	"numpy",
	"opencv-python-headless",
	"pyyaml",
	"scikit-learn",
    "torch",
    "torchvision",
	"tqdm",
	"transformers",
]

[project.urls]
repository = "https://github.com/themattinthehatt/filterworld"
documentation = "https://github.com/themattinthehatt/filterworld"

# project.dependencies are used for metadata when building the project, tool.poetry.dependencies is only used to enrich
# project.dependencies for locking
[tool.poetry.dependencies]
python = ">=3.10,<3.13"

[tool.poetry.scripts]
filterworld = "filterworld.cli.main:main"

[project.optional-dependencies]
dev = [
    "flake8-pyproject",
    "isort",
    "pytest",
    "pytest-html",
]

[tool.isort]
line_length = 99
profile = "black"
src_paths = ["filterworld", "tests"]

[tool.flake8]
max-line-length = 99
extend-ignore = ["E203", "W503", "F722", "W293"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
    "docs/",
    "scripts/",
    "examples/",
]

[tool.pytest.ini_options]
testpaths = "tests"
generate_report_on_test = "True"
markers = ["slow: tests that download models"]
